Use osm_id as feature id for building layer (#827)
Change the feature_id of objects in the tiles from arbitrary number to osm_id.

From discussion here https://github.com/openmaptiles/openmaptiles/pull/725#issuecomment-619059931

Note the multiploygon building are load with negative relation osm_id. But the use in feature_id do an overflow (eg -7980888 -> 18446744073701570000). We can keep as is or not.

No stats done for yet. Trying to let the CI do it.