Simplify landcover layer generalisation SQL
When creating the generalised landcover tables for zooms 9-13, the SQL
was doing something like this:

    SELECT * FROM simplify_vw_z13 WHERE ST_NPoints(geometry) < 50
    UNION ALL
    SELECT * FROM simplify_vw_z13 WHERE ST_NPoints(geometry) >= 50 AND ST_NPoints(geometry) < 300

As in, the simplification/clustering of landcover generalisations was
being done in two steps: once for polygons with fewer than fifty points,
and once for polygons with 50+ points.

As far as I can see, there's no benefit to doing this — it's probably
just an artefact of the dev work done for the pull request
https://github.com/openmaptiles/openmaptiles/pull/1035.

I've simplified the SQL used in the generalisation, merging the two
steps into something like this:

    SELECT * FROM simplify_vw_z13 WHERE ST_NPoints(geometry) < 300

The effect is to slightly reduce the number of landcover features
(polygons with <50 points can now be clustered with those >50), and to
reduce the time taken to generalise the landcover tables (in a small
test using OSM's Iceland data I saw a 7% reduction in the time spent).

Builds upon work from these commits:

https://github.com/openmaptiles/openmaptiles/commit/4a1b0afa26f9b9dcd12f796369bf8c121fd60426
https://github.com/openmaptiles/openmaptiles/commit/da689f9e424dcd2c38aece17ae40b55255a28c4d