Use `docker-compose up -d` for on-going processes
Make targets `start-tileserver`, `start-maputnik`, and `import-osm` initialize on-going processes rather than perform a one-time task. It is more appropriate to execute them using `docker-compose up -d`. With the current `docker-compose run` execution, the user needs to run `make` as a background process. This PR also allows the use of `docker-compose logs` rather than the need to handle `stdout` and `stderr` redirection and avoids the dependency of `stdin` handling for `docker-compose run`

This PR is a follow-up for [a comment](https://github.com/openmaptiles/openmaptiles/commit/0cae7b9fd639f9cbad2a1ce86bbfbcaca4682670#r40912721) on an earlier commit. The comment received a 👍 from @TomPohys.

> According to the [documentation of `imposm run`](https://imposm.org/docs/imposm3/latest/tutorial.html#run), invoked by `import-update`, it runs forever and repeatedly downloads daily/hourly/minutely updates from the server and updates the PostGIS DB. 
IMO, it be more appropriate to `up` it as a docker-compose service rather to `run` it with a shell command.