update_city_point: replace ne.gn_ascii join with ne.name_en (#43)
Here "gn" appears to refer to GeoNames.org as a source. The fields field
`gn_ascii` was deprecated and finally removed in 2021:
https://github.com/nvkelso/natural-earth-vector/issues/477

The deprecation issue recommends to replace the removed field with the
use of `name` (already used) and `name_alts` (which I assume refers to
local variants `name_fr`, `name_en`, ...), so I've included a check on
the English column.